requirements.txt
pillow==8.4.0
streamlit==0.84.1
loguru==0.5.3
tensorflow==2.7
numpy==1.21.2